pyproject.toml
[project]
name = "word-and-zoter-to-latex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypandoc-binary>=1.16.2",
    "pywin32>=306; platform_system == 'Windows'",
    "pyzotero>=1.7.6",
    "tqdm>=4.67.1",
]
